Getting rid of warnings (MSVC phase, II)

+++ blosc/bitshuffle-generic.c
@@ -40,7 +40,7 @@ int64_t bshuf_trans_byte_elem_remainder(const void* in, void* out, const size_t 
             }
         }
     }
-    return size * elem_size;
+    return (int64_t)size * (int64_t)elem_size;
 }
 
 
@@ -68,7 +68,7 @@ int64_t bshuf_trans_bit_byte_remainder(const void* in, void* out, const size_t s
     uint64_t e=1;
     const int little_endian = *(uint8_t *) &e == 1;
     const size_t bit_row_skip = little_endian ? nbyte_bitrow : -nbyte_bitrow;
-    const int64_t bit_row_offset = little_endian ? 0 : 7 * nbyte_bitrow;
+    const size_t bit_row_offset = little_endian ? 0 : 7 * nbyte_bitrow;
 
     CHECK_MULT_EIGHT(nbyte);
     CHECK_MULT_EIGHT(start_byte);
@@ -85,7 +85,7 @@ int64_t bshuf_trans_bit_byte_remainder(const void* in, void* out, const size_t s
             x = x >> 8;
         }
     }
-    return size * elem_size;
+    return (int64_t)size * (int64_t)elem_size;
 }
 
 
@@ -110,7 +110,7 @@ int64_t bshuf_trans_elem(const void* in, void* out, const size_t lda,
                    &in_b[(ii*ldb + jj) * elem_size], elem_size);
         }
     }
-    return lda * ldb * elem_size;
+    return (int64_t)lda * (int64_t)ldb * (int64_t)elem_size;
 }
 
 
@@ -164,7 +164,7 @@ int64_t bshuf_trans_byte_bitrow_scal(const void* in, void* out, const size_t siz
             }
         }
     }
-    return size * elem_size;
+    return (int64_t)size * (int64_t)elem_size;
 }
 
 
@@ -181,7 +181,7 @@ int64_t bshuf_shuffle_bit_eightelem_scal(const void* in, void* out, \
     uint64_t e=1;
     const int little_endian = *(uint8_t *) &e == 1;
     const size_t elem_skip = little_endian ? elem_size : -elem_size;
-    const uint64_t elem_offset = little_endian ? 0 : 7 * elem_size;
+    const size_t elem_offset = little_endian ? 0 : 7 * elem_size;
 
     CHECK_MULT_EIGHT(size);
 
@@ -205,7 +205,7 @@ int64_t bshuf_shuffle_bit_eightelem_scal(const void* in, void* out, \
             }
         }
     }
-    return size * elem_size;
+    return (int64_t)size * (int64_t)elem_size;
 }
